Replace files with those from flappy_Game-main
