--- flappy_Game-main/.gitignore
@@ -1,1 +1,0 @@
-/target
