--- flappy_Game-main/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "flappy"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bracket-lib = "~0.8.7"
-image = "0.24.5"
-rodio = "0.17"
