--- flappy_Game-main/README.md
@@ -1,21 +1,0 @@
-<p align="center">
-<img src="assets/logo.png" width="240">
-</p>
-
-<br />
-<br />
-
-# flappy_Game
-
-![stable-release](https://img.shields.io/badge/flappy_Game-实验性项目-da282a)
-![starts](https://shields.io/github/stars/HighValyrian/flappy_Game)
-
-Welcome to Flappy Dragon！
-
-## content
-
-
-<p align="center">
-<img src="assets/end.png" width="240">
-</p>
-<br/>
